+++ config.php
@@ -3,11 +3,13 @@
 return array (
   //WebIM Config
   'version' => '@VERSION',
+  'debug' => false,
   'isopen' => true,
   'debug' => false,
   'domain' => 'localhost',
   'apikey' => 'public',
-  'server' => 't.nextalk.im:8000',
+  'host' => 't.nextalk.im',
+  'port'    => '8080',
   'emot' => 'default',
   'theme' => 'base',
   'opacity' => '80',